# ARROW-11913: [Rust] Improve performance of StringBuilder by delaying bitmap creation