--- a/rust/arrow/benches/builder.rs
+++ b/rust/arrow/benches/builder.rs
@@ -51,6 +51,20 @@ fn bench_primitive(c: &mut Criterion) {
     group.finish();
 }
 
+fn bench_primitive_nulls(c: &mut Criterion) {
+    let mut group = c.benchmark_group("bench_primitive_nulls");
+    group.bench_function("bench_primitive_nulls", |b| {
+        b.iter(|| {
+            let mut builder = UInt8Builder::new(64);
+            for _ in 0..NUM_BATCHES*BATCH_SIZE {
+                let _ = black_box(builder.append_null());
+            }
+            black_box(builder.finish());
+        })
+    });
+    group.finish();
+}
+
 fn bench_bool(c: &mut Criterion) {
     let data: Vec<bool> = seedable_rng()
         .sample_iter(&Standard)
@@ -92,5 +106,5 @@ fn bench_string(c: &mut Criterion) {
     group.finish();
 }
 
-criterion_group!(benches, bench_primitive, bench_bool, bench_string);
+criterion_group!(benches, bench_primitive, bench_primitive_nulls, bench_bool, bench_string);
 criterion_main!(benches);