@Dandandan 2021-03-09
Thanks a lot @ilya-biryukov the builders are not so efficient at the moment indeed.

I think the future directlion of creating data seems to go more in the direction of using iterators / collect() for creating arrow arrays.

Alternatively I think we could have non-null builders which would not create a bitmap and miss the append_null method.