@ilya-biryukov 2021-03-11
@Dandandan if the input data is generated from existing Rust datastructures/code, using iterators and collect actually does make a lot of sense. 

I believe our use-case will still be centered around the builders. We're looking at speeding up some `datafusion` operations and an extra round-trip through Rust representation would slow things down. Building with iterators looks hard at times, e.g. efficiently constructing an array of strings is actually much simpler with `StringBuilder` than directly in Rust, as `String` allocates on the heap and this can add up when the Strings are many.

> Alternatively I think we could have non-null builders which would not create a bitmap and leave the append_null method in that case.
Actually I would argue that the current approach makes a lot of sense. Nulls don't add that much runtime overhead and adding new buidlers will probably add up to the complexity. Applying the trick from this commit for builders would bring performance really close to optimal, up to a point that this should almost never become a problem in practice.
Note that `PrimitiveBuilder<u8>` is a somewhat pathological case when used inside `StringBuilder`. Paying for bit-set operation on each byte inside the every added string is just too expensive.