@Dandandan 2021-03-11
@ilya-biryukov 
Thanks for updating it with a benchmark. I think this PR is good to go for me as it seems a net win based on the benchmark results. Can you solve the clippy lint?

I am not sure I follow you 100% on using the iterator to build an arrow string array. They are implemented for `&str` as well, so they don't copy the data and should instead translate into a more simple loop. Additionaly, they can pre-allocate the indices array if using an iterator with known size, so they are/should be more efficient than the string builder.
I think we should have a look at the cases where DataFusion is currently using the builder and if possible translate it to use the iterator API.