@ilya-biryukov 2021-03-11
@Dandandan 

> @ilya-biryukov
> Thanks for updating it with a benchmark. I think this PR is good to go for me as it seems a net win based on the benchmark results. Can you solve the clippy lint?

Done.

> I am not sure I follow you 100% on using the iterator to build an arrow string array. They are implemented for `&str` as well, so they don't copy the data and should instead translate into a more simple loop.

The use-case I tried to describe involves [building arrays of unknown length](https://github.com/cube-js/arrow/blob/40e2f0e7e596c71ec9016fef9842e2fbff1ff125/rust/datafusion/src/physical_plan/sorted_aggregate.rs#L23) from multiple inputs that get [sent in batches](https://github.com/cube-js/arrow/blob/40e2f0e7e596c71ec9016fef9842e2fbff1ff125/rust/datafusion/src/physical_plan/sorted_aggregate.rs#L69).


I can hardly see anything more convenient than builders here. But this is definitely not the typical use-case, iterator-based APIs are more convenient in most cases.

> Additionaly, they can pre-allocate the indices array if using an iterator with known size, so they are/should be more efficient than the string builder.
> I think we should have a look at the cases where DataFusion is currently using the builder and if possible translate it to use the iterator API.

Note that the code I linked is **not** upstream  DataFusion, it's from our fork. I should have mentioned that explicitly when mentioning DataFusion in the first place to avoid potential confusion.