@Dandandan 2021-03-11
Ah, makes sense @ilya-biryukov . I remember seeing cube-js before - looks cool.

How is the forking process working for you, do you keep it updated sometimes?
Are there any changes/features we could upstream to DataFusion (what is the license of those changes?)?