@Dandandan 2021-03-11
@jorgecarleitao this seems like a nice performance improvement for the builders to lazily allocate the null bitmap. Should also bring some overall performance improvements to DataFusion or readers that use the builder for non-null columns.
What do you think?